\name{rareNMtests-internal} 
\alias{area.xypolygon}
\title{ Internal rareNMtests function }

\description{ Internal rareNMtests function. }

\usage{
   area.xypolygon(polly)
}

\arguments{
  \item{polly}{A polygon data set.}
}

\value{ The area of the polygon is returned. }

\note{
   The function \code{area.xypolygon} is similar to the \code{area.xypolygon} internal function of the \pkg{spatstat} package.
}

\details{
  This internal function is not usually called directly by the user. 
}

\keyword{ internal }
